Refactored scriptblock manipulation and logic
Rewrote export function to Export-PackageList, which only exports packages to the database.
Refactored how scriptblocks are read in; now they are read in as noteproperties and only converted to a scriptblock when being Invoked. This reduces the number of times scriptblocks have to be converted from string to script.
Updated .format.ps1xml to correctly display scriptblock strings instead.
Fixed Remove-PMPackage to work with removing the last package left.
Updated tests to work with new scriptblock logic.
Added more dummy packages for scriptblock testing,